+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Update postmark template config with new weekly stats templates
     * Deploy new postmark weekly stats templates
     * Update environment with `transition_payment_plans` action
+  * [FEATURE]: Added updating of lineItems to upsertCampaignActiveLineItems
 
 ## v1.6.0 (August 11, 2016)
 * *[v1.6.0-rc1]*